+++ lib/bloc/navigation/navigation_event.dart
@@ -1,0 +1,7 @@
+abstract class NavigationEvent {}
+
+class NavigateToPage extends NavigationEvent {
+  final int index;
+
+  NavigateToPage(this.index);
+}
