+++ lib/bloc/navigation/navigation_state.dart
@@ -1,0 +1,5 @@
+class NavigationState {
+  final int selectedIndex;
+
+  NavigationState(this.selectedIndex);
+}
